feat:new Nail Today, used by the TodayUnit in Title, style tune as well.
