--- app/Component/Blocks/SimpleBlock/styles.module.css
@@ -1,19 +1,0 @@
-.frameNail {
-  display: inline-block;
-  width: 30.8%;
-  min-width: 264px;
-  height: 32vh;
-  min-height: 15.7rem;
-  position: relative;
-  box-sizing: border-box;
-  margin: 0 0 47px;
-  box-shadow: 0 1px 1px -0.1rem, 0rem -0.05rem 2px -0.2rem;
-  border-radius: 0.3rem;
-}
-
-/* beneath must after the main 'frameNail'*/
-
-.frameTwo {
-  width: 48.2%;
-  height: 20.4rem;
-}
