--- app/Component/Nails/NailBasic/styleNailBasic.module.css
@@ -1,35 +1,0 @@
-.Nails_Basic {
-  width: 100%;
-  height: 100%;
-  position: absolute;
-  box-sizing: border-box;
-}
-
-.frameOnMouse {
-  outline: double 1px #777777;
-}
-
-.fontContent {
-  font-size: 1.45rem;
-  font-weight: 400;
-  letter-spacing: 0.07rem;
-  line-height: 1.74rem;
-  font-family: 'gill-sans-nova', 'Noto Sans TC','Lato', sans-serif;
-  word-break: break-all;
-  color: #FAFAFA;
-}
-
-.fontNouns {
-  font-size: 2.14rem;
-  font-weight: 400;
-  letter-spacing: 0.05rem;
-  font-family: 'gill-sans-nova', 'Noto Sans TC','Lato', sans-serif;
-  color: #FAFAFA;
-}
-
-.fontAccount {
-  font-size: 1.3rem;
-  font-weight: 400;
-  letter-spacing: 0.04rem;
-  color: #F0F0F0;
-}
