+++ app/Component/Nails/NailToday/NailToday.jsx
@@ -1,0 +1,165 @@
+import React from 'react';
+import {
+  Link,
+  withRouter
+} from 'react-router-dom';
+import {connect} from "react-redux";
+import classnames from 'classnames';
+import ImgPreview from '../../ImgPreview.jsx';
+import DisplayMarkPreview from '../../Draft/DisplayMarkPreview.jsx';
+import styles from "./styles.module.css";
+import {
+  renderNodesRows
+} from '../utils.js';
+
+class NailToday extends React.Component {
+  constructor(props){
+    super(props);
+    this.state = {
+      onFrame: false,
+      onImg: false
+    };
+    this.nailImgBox = React.createRef();
+    this.nailUnitLink = React.createRef();
+    this._handleEnter_nailFrame = this._handleEnter_nailFrame.bind(this);
+    this._handleLeave_nailFrame = this._handleLeave_nailFrame.bind(this);
+    this._handleEnter_nailImg = this._handleEnter_nailImg.bind(this);
+    this._handleLeave_nailImg = this._handleLeave_nailImg.bind(this);
+    this._render_nails_Marks = this._render_nails_Marks.bind(this);
+    this._render_nails_nouns = this._render_nails_nouns.bind(this);
+    this.style={
+
+    }
+  }
+
+  _handleEnter_nailFrame(e){
+    this.setState({onFrame: true})
+  }
+
+  _handleLeave_nailFrame(e){
+    this.setState({onFrame: false})
+  }
+
+  _handleEnter_nailImg(e){
+    this.setState({onImg: true})
+  }
+
+  _handleLeave_nailImg(e){
+    this.setState({onImg: false})
+  }
+
+  _render_nails_Marks(){
+    let list = this.props.unitBasic.marksList;
+    let marksDOM = [];
+    const self = this;
+
+    for(let i=0 ; i< list.length && i< 3; i++){
+      let key = list[i]
+      marksDOM.push(
+        <div
+          key={"key_nailcosmic_"+self.props.unitId+"_marks_"+i}
+          className={classnames(styles.boxMark, 'fontNailMark', styles.fontMark)}>
+          <DisplayMarkPreview
+            rawContent={self.props.marksBasic[key].editorContent}/>
+        </div>
+      )
+    }
+    return marksDOM;
+  }
+
+  _render_nails_nouns(){
+    /*
+    in order to fit the requirement of common used renderNodesRows(), reduce the modification to existed Nails,
+    we create a props obj, include all the props the renderNodesRows needed,
+    and a nounsList 'exclude' the corner of the day.
+    */
+    let cornerList = this.props.unitBasic.nounsList.filter((node, index)=>{
+      return node != this.props.todayNode
+    });
+    let nounsProps = {
+      unitBasic: {nounsList: cornerList},
+      unitId: this.props.unitId,
+      nounsBasic: this.props.nounsBasic
+    }
+    let nodesDOM = renderNodesRows(nounsProps, styles);
+
+    return nodesDOM;
+  }
+
+  componentDidMount() {
+
+  }
+
+  componentWillUnmount() {
+
+  }
+
+  render(){
+    return(
+      <div
+        className={classnames(
+          styles.frame,
+          {[styles.frameOnMouse]: this.state.onFrame}
+        )}
+        onMouseEnter={this._handleEnter_nailFrame}
+        onMouseLeave={this._handleLeave_nailFrame}>
+        <Link
+          ref={this.nailUnitLink}
+          to={{
+            pathname: this.props.linkPath,
+            search: '?theater&unitId='+this.props.unitId,
+            state: {from: this.props.location}
+          }}
+          className={classnames('plainLinkButton', styles.frame)}
+          style={{display: 'block'}}>
+          <div
+            ref={this.nailImgBox}
+            className={styles.boxImg}
+            onMouseEnter={this._handleEnter_nailImg}
+            onMouseLeave={this._handleLeave_nailImg}>
+            <ImgPreview
+              blockName={''}
+              previewSrc={'/router/img/'+this.props.unitBasic.pic_layer0+'?type=thumb'}
+              _handleClick_ImgPreview_preview={()=>{this.nailImgBox.current.click()}}/>
+            <div
+              className={classnames(
+                styles.boxMask,
+                {[styles.interMask]: this.state.onImg}
+              )}/>
+          </div>
+          <div
+            className={classnames(styles.boxContent)}>
+            <div
+              className={classnames(styles.boxNodes)}>
+              {this._render_nails_nouns()}
+            </div>
+            <div className={styles.boxAuthor}>
+              <span
+                className={classnames('fontNailAuthor', styles.spanAuthor, styles.fontAuthor)}>
+                {this.props.unitBasic.authorId in this.props.usersBasic ? this.props.usersBasic[this.props.unitBasic.authorId].account:null}
+              </span>
+            </div>
+            <div
+              className={classnames(styles.boxMarkPreview)}>
+              {this._render_nails_Marks()}
+            </div>
+          </div>
+        </Link>
+      </div>
+    )
+  }
+}
+
+const mapStateToProps = (state)=>{
+  return {
+    userInfo: state.userInfo,
+    unitCurrent: state.unitCurrent,
+    nounsBasic: state.nounsBasic,
+    usersBasic: state.usersBasic
+  }
+}
+
+export default withRouter(connect(
+  mapStateToProps,
+  null
+)(NailToday));
