+++ app/Component/Nails/NailToday/styles.module.css
@@ -1,0 +1,106 @@
+.frame {
+  width: 100%;
+  height: 100%;
+}
+
+.frameOnMouse {
+  outline: double 1.5px #777777;
+}
+
+.boxImg {
+  width: 66.66%; /* 2/3 */
+  height: 100%;
+  position: absolute;
+  top: 0;
+  box-sizing: border-box;
+}
+
+.boxMask {
+  width: 100%;
+  height: 100%;
+  position:absolute;
+  top:0;
+  left: 0;
+}
+
+.interMask { /*remember must after the boxMask*/
+  width: 100%;
+  height: 100%;
+  position: absolute;
+  background-color: rgba(0,0,0,0.1)
+}
+
+.boxNodes {
+  display: flex;
+  flex-direction: column;
+  width: 90%;
+  max-height: 54%;
+  box-sizing: border-box;
+  padding: 0 4.5% 4% 7%;
+  margin: 7% 0;
+  border-bottom: 1px solid #ff7a5f;
+  overflow: hidden;
+}
+
+.boxNodesItem {
+  /*usually the font were controlled by class in common file of the belonged page.
+  this is for local modified*/
+}
+
+.fontNode {
+  /*usually the font were controlled by class in common file of the belonged page.
+  this is for local modified*/
+  color: #000000;
+}
+
+.boxContent {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  width: 33.33%; /* 1/3 */
+  height: 100%;
+  position: absolute;
+  bottom: 0;
+  right: 0;
+  box-sizing: border-box;
+  padding-left: 4%;
+}
+
+.boxAuthor {
+  max-width: 75%;
+  box-sizing: border-box;
+  padding: 0 0 4% 7%;
+}
+
+.spanAuthor {
+  display: inline-block;
+  position: relative;
+  right: 0;
+  bottom: 0;
+}
+
+.fontAuthor {
+  color: #6e6e6e;
+}
+
+.boxMarkPreview {
+  width: 90%;
+  max-height: 38%;
+  box-sizing: border-box;
+  padding: 0% 6.25%;
+  margin-bottom: 8%;
+  overflow: hidden;
+}
+
+.boxMark {
+  width: 100%;
+  position: relative;
+  box-sizing: border-box;
+  margin-bottom: 1.5rem; /* follow the line height of mark*/
+}
+
+.fontMark {
+  /*usually the font were controlled by class in common file of the belonged page.
+  this is for local modified*/
+  color: #000000;
+}
