+++ app/Within/pageFocus/Main/MainList/MainList.jsx
@@ -161,8 +161,8 @@ class MainList extends React.Component {
   }
 
   _render_IndexNails(){
-    const ruleFirstBlock = [1,1,1,3,3,0,2,1,1,1],
-          ruleBlock = [1,1,1,[0,2],[2,0],1,1,1];
+    const ruleFirstBlock = [2,2,2,3,3,0,2,2,2,2],
+          ruleBlock = [2,2,2,[0,2],[2,0],2,2,2];
 
     let nailsIndex = []; //don't use .map() because we probably need to push twice in one round
     this.props.indexLists.listFocus.forEach((unitId, index)=>{
