+++ app/Within/pageFocus/Main/styles.module.css
@@ -64,8 +64,17 @@
   box-shadow: 0 2px 1px rgb(230, 230, 230);
 }
 
+.boxThree {
+  width: 37.5vw;
+  box-shadow: 1px 2px 1px rgb(230, 230, 230);
+}
+
 .heightNine {
   height: 30vw;
+}
+
+.heightThird {
+  height: 21vw;
 }
 
 .boxFillHoriz {
