+++ public/css/styleWithinCommon.css
@@ -22,10 +22,10 @@
 }
 
 .fontNailMark {
-  font-size: 1.4rem;
+  font-size: 1.34rem;
   font-weight: 400;
-  letter-spacing: 0.034rem;
-  line-height: 17px;
+  letter-spacing: 0.03rem;
+  line-height: 18px;
   text-align: left;
 }
